requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
playwright>=1.41.0
playwright-stealth>=1.0.6
eth-account>=0.10.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0.0
